+++ documentation/cli/change_log.rst
@@ -4,7 +4,7 @@
 FlexMeasures CLI Changelog
 **********************
 
-since v0.12.0 | November XX, 2022
+since v0.12.0 | January 04, 2023
 =================================
 
 * Add ``--resolution``, ``--timezone`` and ``--to-file`` options to ``flexmeasures show beliefs``, to show beliefs data in a custom resolution and/or timezone, and also to save shown beliefs data to a CSV file.
@@ -12,6 +12,8 @@ since v0.12.0 | November XX, 2022
 * Fix ``flexmeasures db-ops dump`` and ``flexmeasures db-ops restore`` incorrectly reporting a success when `pg_dump` and `pg_restore` are not installed.
 * Add ``flexmeasures monitor last-seen``. 
 * Rename ``flexmeasures monitor tasks`` to ``flexmeasures monitor last-run``. 
+* Rename ``flexmeasures add schedule`` to ``flexmeasures add schedule for-storage`` (in expectation of more scheduling commands, based on in-built flex models). 
+
 
 since v0.11.0 | August 28, 2022
 ==============================
@@ -24,7 +26,6 @@ since v0.11.0 | August 28, 2022
 
 since v0.9.0 | March 25, 2022
 ==============================
-
 
 * Add CLI commands for showing data ``flexmeasures show accounts``, ``flexmeasures show account``, ``flexmeasures show roles``, ``flexmeasures show asset-types``, ``flexmeasures show asset``, ``flexmeasures show data-sources``, and ``flexmeasures show beliefs``.
 * Add ``flexmeasures db-ops resample-data`` CLI command to resample sensor data to a different resolution.